cognitive_app/requirements.txt
neo4j==4.4.3
pydantic
spacy
python-dotenv
